+++ driven/awsstorage/adapter.go
@@ -15,6 +15,7 @@
 package awsstorage
 
 import (
+	"bytes"
 	"content/core/model"
 	"fmt"
 	"io"
@@ -136,6 +137,72 @@ func (a *Adapter) DeleteProfileImage(path string) error {
 	_, err = session.DeleteObject(&s3.DeleteObjectInput{
 		Bucket: &a.config.S3ProfileImagesBucket,
 		Key:    &path,
+	})
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// CreateUserVoiceRecord uploads a voice record for the user
+func (a *Adapter) CreateUserVoiceRecord(fileContent []byte, accountID string) (*string, error) {
+	log.Println("Create user voice record")
+
+	s, err := a.createS3Session()
+	if err != nil {
+		log.Printf("Could not create S3 session")
+		return nil, err
+	}
+	key := fmt.Sprintf("names-records/%s.m4a", accountID)
+	objectLocation, err := a.uploadFileToS3(s, bytes.NewReader(fileContent), a.config.S3UsersAudiosBucket, key, "private")
+	if err != nil {
+		log.Printf("Could not upload file")
+		return nil, err
+	}
+
+	return &objectLocation, nil
+}
+
+// LoadUserVoiceRecord loads the voice record for the user
+func (a *Adapter) LoadUserVoiceRecord(accountID string) ([]byte, error) {
+	s, err := a.createS3Session()
+	if err != nil {
+		log.Printf("Could not create S3 session")
+		return nil, err
+	}
+
+	buffer := aws.NewWriteAtBuffer([]byte{})
+
+	key := fmt.Sprintf("names-records/%s.m4a", accountID)
+
+	downloader := s3manager.NewDownloader(s)
+	_, err = downloader.Download(buffer,
+		&s3.GetObjectInput{
+			Bucket: aws.String(a.config.S3UsersAudiosBucket),
+			Key:    aws.String(key),
+		})
+	if err != nil {
+		return nil, err
+	}
+
+	return buffer.Bytes(), nil
+}
+
+// DeleteUserVoiceRecord deletes the voice record for the user
+func (a *Adapter) DeleteUserVoiceRecord(accountID string) error {
+	s, err := a.createS3Session()
+	if err != nil {
+		log.Printf("Could not create S3 session")
+		return err
+	}
+
+	key := fmt.Sprintf("names-records/%s.m4a", accountID)
+
+	session := s3.New(s)
+	_, err = session.DeleteObject(&s3.DeleteObjectInput{
+		Bucket: aws.String(a.config.S3UsersAudiosBucket),
+		Key:    aws.String(key),
 	})
 	if err != nil {
 		return err