+++ driver/web/adapter.go
@@ -93,6 +93,10 @@ func (we Adapter) Start() {
 	contentRouter.HandleFunc("/profile_photo", we.coreAuthWrapFunc(we.apisHandler.StoreProfilePhoto, we.auth.coreAuth.userAuth)).Methods("POST")
 	contentRouter.HandleFunc("/profile_photo", we.coreAuthWrapFunc(we.apisHandler.DeleteProfilePhoto, we.auth.coreAuth.userAuth)).Methods("DELETE")
 
+	contentRouter.HandleFunc("/voice_record", we.coreAuthWrapFunc(we.apisHandler.StoreVoiceRecord, we.auth.coreAuth.userAuth)).Methods("POST")
+	contentRouter.HandleFunc("/voice_record", we.coreAuthWrapFunc(we.apisHandler.GetVoiceRecord, we.auth.coreAuth.userAuth)).Methods("GET")
+	contentRouter.HandleFunc("/voice_record", we.coreAuthWrapFunc(we.apisHandler.DeleteVoiceRecord, we.auth.coreAuth.userAuth)).Methods("DELETE")
+
 	// handle student guide client apis
 	contentRouter.HandleFunc("/student_guides", we.coreAuthWrapFunc(we.apisHandler.GetStudentGuides, we.auth.coreAuth.standardAuth)).Methods("GET")
 	contentRouter.HandleFunc("/student_guides/{id}", we.coreAuthWrapFunc(we.apisHandler.GetStudentGuide, we.auth.coreAuth.standardAuth)).Methods("GET")